% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_index.R
\name{psindex}
\alias{psindex}
\title{bulk import function}
\usage{
psindex(sem_model = NULL, data_set = NULL, bandwidth = 0,
  replications = NULL, iterations_bin = NULL, variables = NULL,
  sd_of_perturbation = 0, control_args = NULL, group = NULL,
  meanstructure = NULL)
}
\description{
A function used to calculate fungible parameter estimates.
}
\examples{
psindex()
}
\author{
Jordan L. Prendez, \email{jordanprendez@gmail.com}
}
\keyword{IR}
