% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_index.R
\name{ps_index}
\alias{ps_index}
\title{Parameter Stability Index}
\usage{
ps_index(model = NULL, data_set = NULL, group = NULL,
  RMSEA_pert = 0, meanstructure = NULL, genanneal_max_iters = 100,
  plot_fpe = FALSE, output_long = FALSE, frac_plot = 1,
  iterations_bin = 40000, suppress_message = FALSE,
  control_genSA = NULL, index_method = "rmsea", lower = -10,
  upper = 10, starting_val_MLE = FALSE)
}
\arguments{
\item{model}{user-specified SEM that using \pkg{lavaan} syntax.}

\item{data_set}{data frame of measured variables.}

\item{RMSEA_pert}{numeric. Defines the maximum size of the perturbation in data-model fit (in the scale of RMSEA). For example, a RMSEA_pert = .01 indicates that all estimates will be stored that have a RMSEA value within .01 RMSEA of the MLE RMSEA value.}

\item{genanneal_max_iters}{integer. The maximum number of iterations of the simulated annealing algorithm.}

\item{plot_fpe}{a logical value indicating whether FPEs should be graphed alongside the maximum likelihood estimates. Defaults to \code{FALSE}.}

\item{output_long}{a logical value that indicates whether the FPEs and the MLE should be output in long data format. Defaults to \code{FALSE}}

\item{frac_plot}{the fraction of FPEs that are graphed. Defaults to 1.}

\item{iterations_bin}{numeric. Represents the maximum number of fungible estimates that can be stored. Defaults to 40,000}

\item{control_genSA}{list. used to control the simulated annealing algorithm (see \link[GenSA]{GenSA} documentation)}

\item{index_method}{Character. Specify the index used to calculate FPEs. Either RMSEA or AIC. Defaults to RMSEA}

\item{lower}{Numeric. Lower bound of the parameter search space to be optimized over. Defaults to \code{-10}}

\item{upper}{Numeric. Upper bound of the parameter search space to be optimized over. Defaults to \code{10}}

\item{starting_val_MLE}{a logical value indicating whether the MLEs should be used as starting values for the SA algorithm. Defaults to \code{FALSE}}
}
\description{
A function used to calculate fungible parameter estimates.
}
\examples{
\dontrun{
model <- '
measurement model
ind60 =~ x1 + x2 + x3
dem60 =~ y1 + y2 + y3 + y4
dem65 =~ y5 + y6 + y7 + y8
# regressions
dem60 ~ ind60
dem65 ~ ind60 + dem60
# residual correlations
y1 ~~ y5
y2 ~~ y4 + y6
y3 ~~ y7
y4 ~~ y8
y6 ~~ y8
'
ps_index(model = model, data_set = PoliticalDemocracy,
        RMSEA_pert = .01,
        plot_fpe   = TRUE,
        index_method = "rmsea",
       frac_plot = .3, iterations_bin = 100000,
        control_genSA = list(maxit = 500))
}
}
\author{
Jordan Yee Prendez, \email{jordanyeeprendez@gmail.com}
}
\keyword{FPE,}
\keyword{estimates}
\keyword{fungible}
\keyword{parameter}
